Implement 2048 Simulation
Implements a 2048 simulation. The board is represented using a four
length array of 16 bit integers. Each integer is a bit-packed row where
four bits represents a tile and the value is the log2 of the tile's
value. Helper functions are included to convert between different
representations of the board. This particular representation was done so
that all possible left slides can be precomputed, and a row's integer
value can be used as an index into the array of the moves to perform a
slide in constant time.