Update build system to use CMake and fix race conditions
Sorry for bad separation of concerns with this commit. This commit adds
an official build setup using CMake to compile the manager and worker
binaries as well as the pybind library. It also fixes some race
conditions in the manager/worker setup, where workers would start
initializing and acquiring a lock before the manager got a chance to
finish initializing.