Rework shared memory data structures, create python bindings
The shared memory array for simulator messages has been replaced with a
lock free queue to avoid the penalty of having to constantly lock and
unlock with a mutex. Python bindings have been created that can be used
to grab simulator states from the queue and place responses in the
response array.